Push push push

+++ frontend/src/components/EventDetails.jsx
@@ -1,41 +1,44 @@
-import { useEventsContext } from "../hooks/useEventsContext";
-import formatDistanceToNow from "date-fns/formatDistanceToNow"; // to make the date appear (we used date fns)
+import { useEventsContext } from '../hooks/useEventsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
+// to make the date appear (we used date fns)
+import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const EventDetails = ({ event }) => {
-    const { dispatch } = useEventsContext();
+  const { dispatch } = useEventsContext()
+  const { user } = useAuthContext()
 
-    const handleClick = async () => {
-        const response = await fetch("/api/events/" + event._id, {
-            method: "DELETE",
-        });
-        const json = await response.json();
+  const handleClick = async () => {
 
-        if (response.ok) {
-            dispatch({ type: "DELETE_EVENT", payload: json });
-        }
-    };
+    if (!user) {
+      return 
+    }
+    const response = await fetch('/api/events/' + event._id, {
 
-    return (
-        <div className="event-details">
-            <h4>{event.title}</h4>
-            <p>
-                <strong>Hours: </strong>
-                {event.hours}
-            </p>
-            <p>
-                <strong>load: </strong>
-                {event.load}
-            </p>
-            <p>
-                {formatDistanceToNow(new Date(event.createdAt), {
-                    addsuffix: true,
-                })}
-            </p>
-            <span className="material-symbols-outlined" onClick={handleClick}>
-                delete
-            </span>
-        </div>
-    );
-};
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
+    })
+    const json = await response.json()
 
-export default EventDetails;
+    if (response.ok) {
+      dispatch({type: 'DELETE_EVENT', payload: json})
+    }
+
+
+  }
+  return (
+    <div className="event-details">
+      <h4>{event.title}</h4>
+      <p><strong>Hours: </strong>{event.hours}</p>
+      <p><strong>load: </strong>{event.load}</p>
+      <p>{formatDistanceToNow(new Date(event.createdAt), { addsuffix: true})}</p>
+      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
+
+
+    </div>
+  )
+
+}
+
+export default EventDetails
